+++ CHANGELOG.md
@@ -9,12 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- The `Defined` type which strips out `undefined` from any type `T`.
+- The `Defined` type, which aliases the `NotUndefined` type. This is because double
+negation breaks people's brain. Easier to read this way. Use whichever mfk.
 - The `Nil` type, a union of `null | undefined`.
 - The `NotNil` type, which strips out `null` and `undefined` from any type `T`.
 This is an alias for `NonNullable`. It's shorter, plays well with `es-toolkit`'s
 terminology, and follows `@infra-blocks` naming conventions.
 - The `NotNull` type, which strips out `null` from any type `T`.
+- The `NoUndefined` type which strips out `undefined` from any type `T`.
 
 ### Removed